backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
psutil>=5.9.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
websockets>=12.0
